Generalize languages
This commit makes the languages list be a parameter that can be passed in
when calling the newly exported Run method.
