+++ cmd/gazelle/BUILD.bazel
@@ -3,17 +3,17 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
+        "command.go",
         "diff.go",
         "fix.go",
         "fix-update.go",
-        "gazelle.go",
         "langs.go",
         "print.go",
         "update-repos.go",
         "version.go",
     ],
     importpath = "github.com/bazelbuild/bazel-gazelle/cmd/gazelle",
-    visibility = ["//visibility:private"],
+    visibility = ["//visibility:public"],
     deps = [
         "//internal/config:go_default_library",
         "//internal/flag:go_default_library",
@@ -33,23 +33,25 @@ go_library(
 
 go_binary(
     name = "gazelle",
-    embed = [":go_default_library"],
+    srcs = ["gazelle.go"],
     msan = "off",
     pure = "off",
     race = "off",
     static = "off",
     visibility = ["//visibility:public"],
+    deps = [":go_default_library"],
 )
 
 go_binary(
     name = "gazelle_pure",
-    embed = [":go_default_library"],
+    srcs = ["gazelle.go"],
     msan = "off",
     pure = "on",
     race = "off",
     static = "off",
     tags = ["manual"],
     visibility = ["//visibility:public"],
+    deps = [":go_default_library"],
 )
 
 go_test(
